predict/requirement.txt
nltk
Flask
beautifulsoup4
newspaper3k
